refactor(contest): убрана промежуточная страница урока
Между списком уроков и тренажёром висел лишний экран: видео, описание и
кнопка «Выполнить задание». Он ничего не добавлял — задание, ответ и тот же
ролик уже есть в тренажёре.

- страница /cours/[task] удалена;
- ролик показывается заставкой прямо в карточке списка, смотреть можно
  оттуда;
- «К уроку» ведёт сразу в тренажёр: ставит конкурсный ключ и открывает
  экран задания.

Проверено в браузере: клик по «К уроку» открывает тренажёр на первом уроке
с заданием и полем ответа.
